% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.E3D.R
\name{alpha.E3D}
\alias{alpha.E3D}
\title{calculate Van-Genuchten ALPHA}
\usage{
alpha.E3D(Sa, Cl, Bulk, Corg)
}
\arguments{
\item{Sa}{Gravimetric content of sand particles in fine soil (0-100)}

\item{Cl}{Gravimetric content of clay particles in fine soil (0-100)}

\item{Bulk}{bulk density of dry soil [kg/m^3]}

\item{Corg}{Gravimetric content of organic bound carbon in fine soil (0-100)}
}
\value{
alpha
}
\description{
calculates alpha-parameter according to Vereecken et. al 1989 (doi:10.1097/00010694-198912000-00001)
for Van-Genuchten soil-water retention curve (Van-Genuchten 1980 - doi:10.2136/sssaj1980.03615995004400050002x)
-> first equation in Table 7 for log(alpha)
}
\examples{
alpha.E3D(30,20,1300,1.3)
}
