% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{calc_cumq_cumsed}
\alias{calc_cumq_cumsed}
\title{calculation of cumulative runoff [l] from time [s] and runoff rate [l/s]}
\usage{
calc_cumq_cumsed(timesteps, qsteps)
}
\value{
A numeric vector giving number of characters (code points) in each
   element of the character vector. Missing string have missing length.
}
\description{
Technically this returns the number of "code points", in a string. One
code point usually corresponds to one character, but not always. For example,
an u with a umlaut might be represented as a single character or as the
combination a u and an umlaut.
}
\examples{
str_length(letters)
str_length(NA)
str_length(factor("abc"))
str_length(c("i", "like", "programming", NA))
}
\seealso{
[stringi::stri_length()] which this function wraps.
}
