% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_soilset.R
\name{calc_soilset}
\alias{calc_soilset}
\title{evaluate soil parameters on hillslopes in EROSION-3D}
\usage{
calc_soilset(
  soils = dummy_soilset(),
  intensity = 0.7,
  plotlength = 22,
  slope = 9,
  endmin = 30,
  resolution = 1,
  ponding = FALSE,
  path = tempdir()
)
}
\arguments{
\item{soils}{dataframe, holding all EROSION-3D soil parameters in named columns}

\item{intensity}{numeric vector, rainfall intensity of preceding time interval - corespondeces to endmin}

\item{plotlength}{integer value, length of experimental plot, needs to be an integer due to spatial resolution of 1 meter in E3D}

\item{slope}{integer value, mean slope of experimental plot in percent}

\item{endmin}{numeric vector, duration since start of rainfall experiment in full minutes, length must equal length of intensity}

\item{resolution}{set spatial resoultion}

\item{ponding}{logical TRUE means ponding option is used, FALSE - is not used in E3D, ponding limits amount of infiltrating water to available water}

\item{path}{path to modeling directory, default is a temporary directory}
}
\description{
This function takes a dataframe including all soil parameters of EROSION-3D and models one hillslope for each of given parameter sets.
It returns a dataframe with POLY_ID as identifier and calculated cumulative runoff [litres] and soil loss [kg].
}
\examples{
calc_soilset(soil_params)

}
