% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_settings.E3D.R
\name{change_settings.E3D}
\alias{change_settings.E3D}
\title{change model settings for E3D}
\usage{
change_settings.E3D(path = NA, filename = "model/run.par",
  module = NA, setting = NA, value = NA, setpath = TRUE)
}
\arguments{
\item{path}{folder path in which file will be written}

\item{filename}{name of created file, should be *.par}

\item{module}{character vector, defines sub module in which setting will be changed}

\item{setting}{character vector, defines setting to be changed}

\item{value}{character vector, defines value to be changed to}

\item{setpath}{boolean, shortcut to set working directory for E3D - if TRUE the path will be written as project path for rain, dem, soil and result}
}
\value{
none, *.par file is written
}
\description{
This function changes the settings for E3D models in the *.par file
}
