% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_setup.E3D.R
\name{check_setup.E3D}
\alias{check_setup.E3D}
\title{check e3d setup}
\usage{
check_setup.E3D()
}
\value{
numeric, exit code of command line
}
\description{
checks, if e3d can be run from command line
}
