% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.E3D.R
\name{create_folders.E3D}
\alias{create_folders.E3D}
\title{creates folder structure for E3D modelling}
\usage{
create_folders.E3D(path, dummy_relief.log = FALSE)
}
\value{
creates folders
}
\description{
creates folder structure for E3D modelling
}
\examples{

}
