% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.LIS.R
\name{create_folders.LIS}
\alias{create_folders.LIS}
\title{creates folder structure including a basic openLISEM model}
\usage{
create_folders.LIS(path = tempdir(), overwrite = FALSE)
}
\arguments{
\item{path}{folder path in which file will be written}

\item{overwrite}{boolean, TRUE allows overwriting of files}
}
\value{
file path to created folder structure
}
\description{
creates folder structure including a basic openLISEM model
}
\examples{
create_folders.LIS(path = "C:/LISmodel/", overwrite = TRUE)
}
