% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine.eros.cumsed.E3D.R
\name{determine.eros.cumsed.E3D}
\alias{determine.eros.cumsed.E3D}
\title{resistance to erosion determination}
\usage{
determine.eros.cumsed.E3D(
  FCl,
  MCl,
  CCl,
  FSi,
  MSi,
  CSi,
  FSa,
  MSa,
  CSa,
  Corg,
  Bulk,
  Moist,
  Skin,
  Roughness,
  Cover,
  Soilloss,
  intensity,
  plotwidth,
  plotlength,
  slope,
  endmin,
  ponding = FALSE,
  simlines = 100,
  path =
    "C:\\\\Users\\\\Jonas.Lenz\\\\Desktop\\\\e3d-dev\\\\e3dwebservice\\\\e3dtestdata\\\\skin",
  silent = TRUE,
  resolution = 1,
  errorcode = FALSE
)
}
\arguments{
\item{FCl}{numeric}

\item{MCl}{numeric}

\item{CCl}{numeric}

\item{FSi}{numeric}

\item{MSi}{numeric}

\item{CSi}{numeric}

\item{FSa}{numeric}

\item{MSa}{numeric}

\item{CSa}{numeric}

\item{Corg}{numeric}

\item{Bulk}{numeric}

\item{Moist}{numeric}

\item{Skin}{numeric}

\item{Roughness}{numeric}

\item{Cover}{numeric}

\item{Soilloss}{numeric value, cumulative sediment loss in kg is the fitting target of this function}

\item{intensity}{numeric vector, rainfall intensity of preceding time interval - correspondences to endmin}

\item{plotwidth}{numeric value, width of the experimental plot, CumRunoff will be normalized to one meter width using this parameter}

\item{plotlength}{numeric value, length of experimental plot, needs to be an multiple of spatial resolution (default 1 meter)}

\item{slope}{integer value, mean slope of experimental plot in percent}

\item{endmin}{numeric vector, duration since start of rainfall experiment in full minutes, length must equal length of intensity}

\item{ponding}{logical TRUE means ponding option is used, FALSE - is not used in E3D, ponding limits amount of infiltrating water to available water}

\item{simlines}{integer value, number of parallel calculated plots, higher numbers decrease number of iteration steps with E3D, but increases number of write-read operations}

\item{path}{path to modeling directory, default is a temporary directory}

\item{silent}{logical, if TRUE skinfactor iteration steps will be written as message}

\item{resolution}{set spatial resolution [m]}

\item{errorcode}{logical, if TRUE an errorcode is returned instead of NA for experiments, which can not be determined: -1 No runoff, -2 to high sediment concentration}
}
\description{
This function calibrates parameter resistance to erosion that calculated cumulative sediment loss equals the measured value from a rainfall experiment.
Therefore it models a rainfall-experiment plot using EROSION-3D.
}
\examples{
determine.eros.cumsed.E3D(FCl=5,MCl=10,CCl=15, FSi=10,MSi=20,CSi=10, FSa=15,MSa=10,CSa=5, Corg = 1.3, Bulk = 1300, Moist = 22, Skin = 0.005,Roughness=0.05, Cover = 20, Soilloss = 1, intensity = 0.5, plotwidth = 1, plotlength = 10, slope = 10, endmin = 30, ponding = TRUE)
determine.eros.cumsed.E3D(FCl=5,MCl=10,CCl=15, FSi=10,MSi=20,CSi=10, FSa=15,MSa=10,CSa=5, Corg = 1.3, Bulk = 1300, Moist = 22, Skin = 0.005,Roughness=0.05, Cover = 20, Soilloss = 1, intensity = 0.5, plotwidth = 1, plotlength = 10, slope = 10, endmin = 30, ponding = FALSE)

}
