% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_soilset.R
\name{dummy_soilset}
\alias{dummy_soilset}
\title{gets soil params example}
\usage{
dummy_soilset(path = tempdir())
}
\description{
This function returns an example dataframe including all soil parameters of EROSION-3D.
}
\examples{
dummy_soilset(soil_params)

}
