% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch_params.R
\name{e3d_watchcell_get_result}
\alias{e3d_watchcell_get_result}
\title{get watched results of EROSION-3D}
\usage{
e3d_watchcell_get_result(watchcellid, path = tempdir())
}
\arguments{
\item{watchcellid}{integer, id of parameter of interest - explanation and possible valuesin e3d_watchcell_IDs}

\item{path}{path model folder}
}
\description{
This function gets the results of an watched parameter and stacks them in a raster stack. Make sure to set watch method before running e3d by e3d_watchcell_set()!
}
\examples{
e3d_watchcell_set <- function(2, path = tempdir())

}
