% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch_params.R
\name{e3d_watchcell_set}
\alias{e3d_watchcell_set}
\title{watch results of EROSION-3D}
\usage{
e3d_watchcell_set(
  watchcellid,
 
    path_to_ini = "C:\\\\Users\\\\Jonas.Lenz\\\\AppData\\\\Roaming\\\\GeoGnostics\\\\Erosion3D-330\\\\"
)
}
\arguments{
\item{watchcellid}{integer, id of parameter of interest - explanation and possible valuesin e3d_watchcell_IDs}

\item{path_to_ini, }{path to "e3d.ini" in standard instatlation stored in "C:\\Users\\<user>\\AppData\\Roaming\\GeoGnostics\\Erosion3D-330\\"}
}
\description{
This function sets the flag, that a parameter value is recorded in each time step by EROSION-3D
}
\examples{
e3d_watchcell_set <- function(2, path_to_ini = "C:\\\\Users\\\\Jonas.Lenz\\\\AppData\\\\Roaming\\\\GeoGnostics\\\\Erosion3D-330\\\\")

}
