% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version.E3D.R
\name{get_version.E3D}
\alias{get_version.E3D}
\title{reads version of E3D installation}
\usage{
get_version.E3D(build = FALSE, install_path = NA, relief.v = FALSE)
}
\arguments{
\item{build}{logical, TRUE returns a string including build information}

\item{install_path, }{path to installation directory of EROSION-3D}

\item{relief.v}{logical, TRUE returns also version of terrain.dll}
}
\value{
string E3D version number
}
\description{
Returns the version number of the E3D installation given in the relief.log file in relief set created by E3D.
}
\examples{
get_version.E3D()
}
