% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nordpol.E3D.R
\name{nordpol.E3D}
\alias{nordpol.E3D}
\title{calculate Van-Genuchten n}
\usage{
nordpol.E3D(Sa, Cl)
}
\arguments{
\item{Sa}{Gravimetric content of sand particles in fine soil (0-100)}

\item{Cl}{Gravimetric content of clay particles in fine soil (0-100)}
}
\value{
n
}
\description{
calculates n-parameter according to Vereecken et. al 1989 (doi:10.1097/00010694-198912000-00001)
for Van-Genuchten soil-water retention curve (Van-Genuchten 1980 - doi:10.2136/sssaj1980.03615995004400050002x)
-> first equation in Table 7 for log(n)
}
\examples{
nordpol.E3D(30,20)
}
