% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psimo_to_moist.E3D.R
\name{pF_to_moist.E3D}
\alias{pF_to_moist.E3D}
\title{calculate water content to achieve given pF for E3D}
\usage{
pF_to_moist.E3D(THETA_S, THETA_R, pF, ALPHA, NORDPOL, digits = 2)
}
\arguments{
\item{THETA_S}{Maximum volumetric water content (0-100)}

\item{THETA_R}{Minimum volumetric water content (0-100)}

\item{pF}{target pF value}

\item{ALPHA}{Van-Genuchten parameter}

\item{NORDPOL}{Van-Genuchten parameter}

\item{digits}{integer, number of decimal digits}
}
\value{
moist
}
\description{
calculates water content from given pF-value (matrix potential).
a reverse function of psimo.E3D
}
\seealso{
psimo.E3D
}
