% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psimo.E3D.R
\name{psimo.E3D}
\alias{psimo.E3D}
\title{calculate  matrix potential}
\usage{
psimo.E3D(THETA_S, THETA_R, INITMOIST, ALPHA, NORDPOL)
}
\arguments{
\item{THETA_S}{Maximum volumetric water content (0-100)}

\item{THETA_R}{Minimum volumetric water content (0-100)}

\item{INITMOIST}{actual volumetric water content (0-100)}

\item{ALPHA}{Van-Genuchten parameter}

\item{NORDPOL}{Van-Genuchten parameter}
}
\value{
psimo
}
\description{
calculates matrix potential PsimO [hPa] according to Van-Genuchten 1980 - doi:10.2136/sssaj1980.03615995004400050002x
as cited in Schmidt 1996 (ISBN 978-3-88009-062-0)
}
\seealso{
alpha.E3D, nordpol.E3D, theta_s.E3D, theta_r.E3D
}
