% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psimo_to_moist.E3D.R
\name{psimo_to_moist.E3D}
\alias{psimo_to_moist.E3D}
\title{calculate water content for given matrix potential for E3D}
\usage{
psimo_to_moist.E3D(THETA_S, THETA_R, pF, ALPHA, NORDPOL, digits = 2)
}
\value{
moist
}
\description{
calculates water content from given pF-value (matrix potential)
}
\examples{

}
\seealso{
psimo.E3D
}
