% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_result.E3D.R
\name{read_result.E3D}
\alias{read_result.E3D}
\title{read E3D result set}
\usage{
read_result.E3D(result_layer, modelpath = tempdir())
}
\arguments{
\item{result_layer, }{output layer of EROSION-3D}

\item{modelpath, }{path to calculated model folder}
}
\value{
none
}
\description{
Internal function to read result set of E3D
}
