% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pourpoints.R
\name{set_pourpoints}
\alias{set_pourpoints}
\title{watch results of EROSION-3D using pour points}
\usage{
set_pourpoints(
  activate_pour = TRUE,
 
    path_to_ini = "C:\\\\Users\\\\Jonas.Lenz\\\\AppData\\\\Roaming\\\\GeoGnostics\\\\Erosion3D-330\\\\"
)
}
\arguments{
\item{path_to_ini, }{path to "e3d.ini" in standard instatlation stored in "C:\\Users\\<user>\\AppData\\Roaming\\GeoGnostics\\Erosion3D-330\\"}

\item{watchcellid}{integer, id of parameter of interest - explanation and possible valuesin e3d_watchcell_IDs}
}
\description{
This function sets the flag, that a parameter value is recorded in each time step by EROSION-3D
}
\examples{
activate_pourpoints <- function(activate_pour = TRUE, path_to_ini = "C:\\\\Users\\\\Jonas.Lenz\\\\AppData\\\\Roaming\\\\GeoGnostics\\\\Erosion3D-330\\\\")

}
