% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_r.E3D.R
\name{theta_r.E3D}
\alias{theta_r.E3D}
\title{calculate THETA_R}
\usage{
theta_r.E3D(Cl, Corg)
}
\arguments{
\item{Cl}{Gravimetrical content of clay particles in fine soil (0-100)}

\item{Corg}{Gravimetrical content of organic bound carbon in fine soil (0-100)}
}
\value{
THETA_R
}
\description{
calculates THETA_R-parameter (minimum soil moisture) according to Vereecken et. al 1989 (doi:10.1097/00010694-198912000-00001)
-> equation in Table 7
}
\examples{
theta_r.E3D(20,1.3)
}
