% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_s.E3D.R
\name{theta_s.E3D}
\alias{theta_s.E3D}
\title{calculate THETA_S}
\usage{
theta_s.E3D(Cl, Bulk)
}
\value{
THETA_S
}
\description{
calculates THETA_S-parameter (maximum soil moisture) according to Vereecken
}
\examples{

}
