% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.landuse.E3D.R
\name{write.landuse.E3D}
\alias{write.landuse.E3D}
\title{Write an synthetic EROSION-3D land classification input file}
\usage{
write.landuse.E3D(
  POLY_ID = c(1, 2, 3),
  length = 50,
  path,
  filename = "/model/landuse.asc",
  resolution = 1
)
}
\arguments{
\item{POLY_ID}{numeric vector, number of elements defines columns of created relief file. Refers to POLY_ID in soil_params.csv of E3D.}

\item{length}{numeric value, defining slope length in whole meters}

\item{path}{folder path in which file will be written}

\item{filename}{name of created file}

\item{resolution}{set spatial resoultion}
}
\value{
None, file is written
}
\description{
This function writes a land-use/land-cover file for simultaneously modeling of EROSION-2D slopes in EROSION-3D.
One row represents one EROSION-2D slope, when flow routing is set to one neighbor in EROSION-3D.
}
\examples{
write.landuse.E3D(path = "C:/E3Dmodel/")
}
