% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.lulc.E3D.R
\name{write.lulc.E3D}
\alias{write.lulc.E3D}
\title{Write an synthetic EROSION-3D landclassification input file}
\usage{
write.lulc.E3D(POLY_ID = c(1, 2, 3), length = 50, path,
  filename = "lulc.asc")
}
\value{
No return value, but file is written
}
\description{
This function writes a landuse/landcover file for simultaneously modelling of EROSION-2D slopes in EROSION-3D.
One row represents one homogeneous soil parameter set.
Length defines slope length.
}
\examples{
write.lulc.E3D(path = "C:/")
}
