% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pourpoints.R
\name{write.pourpoint.E3D}
\alias{write.pourpoint.E3D}
\title{Write POUR point raster}
\usage{
write.pourpoint.E3D(
  POLY_ID,
  length,
  path,
  filename = "POUR.asc",
  resolution = 1
)
}
\arguments{
\item{POLY_ID}{numeric vector, number of elements defines columns of created relief file. Refers to POLY_ID in soil_params.csv of E3D.}

\item{length}{numeric value, defining slope length in whole meters}

\item{path}{folder path in which file will be written}

\item{filename}{name of created file}

\item{resolution}{set spatial resolution}
}
\value{
None, file is written
}
\description{
Create a pour point raster to watch virtual slopes at outlet (lowest point of each slope)
}
\examples{
write.pourpoint.E3D(POLY_ID = c(1,2,3),length = 50, path = "C:/E3Dmodel/relief/", filename = "dem.asc")
}
