% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.relief.E3D.R
\name{write.relief.E3D}
\alias{write.relief.E3D}
\title{Write an synthetic EROSION-3D relief input file}
\usage{
write.relief.E3D(POLY_ID = c(1, 2, 3), length = 50, slope = 11, path,
  filename = "dem.asc")
}
\value{
No return value, but file is written
}
\description{
This function writes a relief file for simultaneously modelling of EROSION-2D slopes in EROSION-3D.
One row represents one EROSION-2D slope.
Length defines slope length.
Slope is given in %.
}
\examples{
write.relief.E3D(path = "C:/")
}
