% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.E3D.R
\name{alpha.E3D}
\alias{alpha.E3D}
\title{calculate ALPHA}
\usage{
alpha.E3D(Sa, Cl, Bulk, Corg)
}
\value{
alpha
}
\description{
calculates alpha-parameter accordng to Vereecken
}
\examples{
alpha.E3D(30,20,1300,1.3)
}
