% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cumq_cumsed.R
\name{calc_cumq_cumsed}
\alias{calc_cumq_cumsed}
\title{calculation of cumulative runoff [l] from time [s] and runoff rate [l/s]}
\usage{
calc_cumq_cumsed(timesteps, qsteps)
}
\description{
This function calculates cummulative runoff from a timeline with runoff rates.
}
\examples{

}
