% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_setup.E3D.R
\name{check_setup.E3D}
\alias{check_setup.E3D}
\title{check e3d setup}
\usage{
check_setup.E3D()
}
\value{
none
}
\description{
checks, if e3d is available from global %PATH% or can be found in registry.
If registry entry is found, it will be added to PATH for this session.
}
