% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.E3D.R
\name{create_folders.E3D}
\alias{create_folders.E3D}
\title{creates folder structure including a basic E3D model}
\usage{
create_folders.E3D(path = tempdir(), overwrite = FALSE)
}
\arguments{
\item{path}{folder path in which file will be written}

\item{overwrite}{boolean, TRUE allows overwriting of files}
}
\value{
file path to created folder structure
}
\description{
creates folder structure including a basic E3D model
}
\examples{
create_folders.E3D(path = "C:/E3Dmodel/", overwrite = TRUE)
}
