% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine.eros.cumsed.E3D.R
\name{determine.eros.cumsed.E3D}
\alias{determine.eros.cumsed.E3D}
\title{resistance to erosion determination}
\usage{
determine.eros.cumsed.E3D(FCl, MCl, CCl, FSi, MSi, CSi, FSa, MSa, CSa,
  Corg, Bulk, Moist, Skin, Roughness, Cover, Soilloss, intensity,
  plotwidth, plotlength, slope, endmin, ponding = FALSE,
  simlines = 100, path = tempdir(), silent = TRUE)
}
\arguments{
\item{FCl}{numeric}

\item{MCl}{numeric}

\item{CCl}{numeric}

\item{FSi}{numeric}

\item{MSi}{numeric}

\item{CSi}{numeric}

\item{FSa}{numeric}

\item{MSa}{numeric}

\item{CSa}{numeric}

\item{Corg}{numeric}

\item{Bulk}{numeric}

\item{Moist}{numeric}

\item{Skin}{numeric}

\item{Roughness}{numeric}

\item{Cover}{numeric}

\item{Soilloss}{numeric value, cummulative sediment loss in kg is the fitting target of this function}

\item{intensity}{numeric value, a constant rainfall intensity during the experiment}

\item{plotwidth}{numeric value, width of the experimental plot, CumRunoff will be normalized to one meter width using this parameter}

\item{plotlength}{integer value, length of experimental plot, needs to be an interger due to spatial resolution of 1 meter in E3D}

\item{slope}{integer value, mean slope of experimental plot in percent}

\item{endmin}{integer value, duration of rainfall experiment in full minutes rainfall was applied.}

\item{ponding}{logical TRUE means ponding option is used, FALSE - is not used in E3D, ponding limits amount of infiltrating water to available water}

\item{simlines}{integer value, number of parallel computated plots, higher numbers decrease number of iteration steps with E3D, but increases number of write-read operations}

\item{path}{path to modelling directory, defaul is a temporary directory}

\item{silent}{logical, if TRUE skinfactor iteration steps will be written as message}
}
\description{
This function calibrates parameter resistance to erosion that calculated cummulative sediment loss equals the measured value from a rainfall experiment.
Therfore it models a rainfall-experiment plot using EROSION-3D.
}
\examples{
determine.eros.cumsed.E3D(FCl=5,MCl=10,CCl=15, FSi=10,MSi=20,CSi=10, FSa=15,MSa=10,CSa=5, Corg = 1.3, Bulk = 1300, Moist = 22, Skin = 0.005,Roughness=0.05, Cover = 20, Soilloss = 1, intensity = 0.5, plotwidth = 1, plotlength = 10, slope = 10, endmin = 30, ponding = TRUE)
determine.eros.cumsed.E3D(FCl=5,MCl=10,CCl=15, FSi=10,MSi=20,CSi=10, FSa=15,MSa=10,CSa=5, Corg = 1.3, Bulk = 1300, Moist = 22, Skin = 0.005,Roughness=0.05, Cover = 20, Soilloss = 1, intensity = 0.5, plotwidth = 1, plotlength = 10, slope = 10, endmin = 30, ponding = FALSE)

}
