% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine.skin.infil.E3D.R
\name{determine.skin.infil.E3D}
\alias{determine.skin.infil.E3D}
\title{Skinfactor determination}
\usage{
determine.skin.infil.E3D(
  Cl,
  Si,
  Sa,
  Corg,
  Bulk,
  Moist,
  infilrate,
  intensity,
  plotwidth = 1,
  plotlength = 2,
  slope = 5,
  endmin,
  ponding = FALSE,
  simlines = 100,
  path = tempdir(),
  silent = TRUE,
  version = get_version.E3D()
)
}
\arguments{
\item{Cl}{numeric}

\item{Si}{numeric}

\item{Sa}{numeric}

\item{Corg}{numeric}

\item{Bulk}{numeric}

\item{Moist}{numeric}

\item{infilrate}{numeric value, infiltration rate is the fitting target of this function}

\item{intensity}{numeric vector, rainfall intensity of preceding time interval - corespondeces to endmin}

\item{plotwidth}{numeric value, width of the experimental plot, CumRunoff will be normalized to one meter width using this parameter}

\item{plotlength}{integer value, length of experimental plot, needs to be an integer due to spatial resolution of 1 meter in E3D}

\item{slope}{integer value, mean slope of experimental plot in percent}

\item{endmin}{numeric vector, duration since start of rainfall experiment in full minutes, length must equal length of intensity}

\item{ponding}{logical TRUE means ponding option is used, FALSE - is not used in E3D, ponding limits amount of infiltrating water to available water}

\item{simlines}{integer value, number of parallel calculated plots, higher numbers decrease number of iteration steps with E3D, but increases number of write-read operations}

\item{path}{path to modeling directory, default is a temporary directory}

\item{silent}{logical, if TRUE skinfactor iteration steps will be written as message}

\item{version}{version number can be set manually if known to reduce calls to E3D and save processing time}
}
\description{
This function calibrates parameter skinfactor that calculated infiltration rate equals the measured value from a rainfall experiment.
Therefore it models a rainfall-experiment plot using EROSION-3D.
}
\examples{
determine.skin.infil.E3D(Cl = 30, Si = 40, Sa = 30, Corg = 1.3, Bulk = 1300, Moist = 22, infilrate = 0.2, intensity = 0.5, plotwidth = 1, plotlength = 10, slope = 10, endmin = 30, ponding = TRUE, silent = FALSE)

}
