% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psimo.E3D.R
\name{psimo.E3D}
\alias{psimo.E3D}
\title{calculate matrix potential according to Vereecken}
\usage{
psimo.E3D(THETA_S, THETA_R, INITMOIST, ALPHA, NORDPOL)
}
\value{
psimo
}
\description{
calculates matrix potential PsimO [hPa] according to Vereecken
}
\examples{

}
