% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_r.E3D.R
\name{theta_r.E3D}
\alias{theta_r.E3D}
\title{calculate THETA_R}
\usage{
theta_r.E3D(Cl, Corg)
}
\value{
THETA_R
}
\description{
calculates THETA_R-parameter (minimum soil moisture) according to Vereecken
}
\examples{

}
