% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_s.E3D.R
\name{theta_s.E3D}
\alias{theta_s.E3D}
\title{calculate THETA_S}
\usage{
theta_s.E3D(Cl, Bulk)
}
\arguments{
\item{Cl}{Gravimetrical content of clay particles in fine soil (0-100)}

\item{Bulk}{bulkdensity of dry soil [kg/m^3]}
}
\value{
THETA_S
}
\description{
calculates THETA_S-parameter (maximum soil moisture) according to Vereecken et. al 1989 (doi:10.1097/00010694-198912000-00001)
-> equation in Table 7
}
\examples{
theta_s.E3D(20,1300)
}
