% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.rainfile.E3D.R
\name{write.rainfile.E3D}
\alias{write.rainfile.E3D}
\title{Write an EROSION-3D rainintensity input file}
\usage{
write.rainfile.E3D(x = cbind.data.frame(time = c(0, 60, 120, 240), intens
  = c(0.1, 0.6, 1.2, 2.4)), date = "1990-10-27",
  filename = "rain_e3d.reg", timestep = 60)
}
\value{
No return value, but file is written
}
\description{
This function writes a rainintensity input file, which can be used as input in EROSION-3D.
x shall be a dataframe containing a timeline
}
\examples{
write.rainfile.E3D()
}
