% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.relief.E3D.R
\name{write.relief.E3D}
\alias{write.relief.E3D}
\title{Write an synthetic EROSION-3D relief input file}
\usage{
write.relief.E3D(POLY_ID, length, slope, path, filename = "dem.asc")
}
\arguments{
\item{POLY_ID}{numeric vector, number of elements defines columns of created relief file. Refers to POLY_ID in soil_params.csv of E3D.}

\item{length}{numeric value, defining slope length in whole meters}

\item{slope}{numeric value, defining slope inclination in whole percent}

\item{path}{folder path in which file will be written}

\item{filename}{name of created file}
}
\value{
None, file is written
}
\description{
This function writes a relief file for simultaneously modelling of independent hillslopes (EROSION-2D) in EROSION-3D.
One row represents one EROSION-2D slope, when flow routing is set to one neighbour in EROSION-3D.
}
\examples{
write.relief.E3D(POLY_ID = c(1,2,3),length = 50,slope = 11, path = "C:/E3Dmodel/", filename = "dem.asc")
}
